fix: Resolve ALL build errors - complete pass
- DiagnosticsDashboardSection: remove LocalContext param (use val inside)
- ProductionPipelineSection: remove LocalContext param (use val inside)
- DeveloperDashboardScreen: update callers to new signatures
- SettingsScreen: add missing closing brace for CollapsibleSettingsCard #3
- MainActivity: move handleDeepLink/onNewIntent inside class, fix Color cast
- ProcessingScreen: hoist finalScenes to composable level, fix Icons.Default.Send
- AppDiagnostics: fix Long->Int cast for totalCalls

+++ app/src/main/java/com/qabas/app/AppDiagnostics.kt
@@ -334,7 +334,7 @@ object AppDiagnostics {
                 isConfigured = isConfigured,
                 lastLatencyMs = apiStat?.lastLatencyMs ?: 0,
                 successRate = apiStat?.successRate ?: 0f,
-                totalCalls = apiStat?.totalCalls ?: 0,
+                totalCalls = (apiStat?.totalCalls ?: 0L).toInt(),
                 status = status
             )
         }
